move Auth types to a better location
